refactor: 정수 타입을 실수 타입으로 변경
- Integer를 Double 타입으로 변경
- PositiveNumberExtractor 리스트 타입 Double로 변경
- Calculator Computer OutputView 모두 타입 Double로 변경

+++ src/main/java/calculator/Computer.java
@@ -18,8 +18,8 @@ public class Computer {
 
     public void execute() {
         String inputString = inputView.input();
-        List<Integer> numbers = positiveNumberExtractor.extract(inputString);
-        Integer sum = calculator.sum(numbers);
+        List<Double> numbers = positiveNumberExtractor.extract(inputString);
+        Double sum = calculator.sum(numbers);
         outputView.printResult(sum);
     }
 }
